+++ packages/wenyou_api/lib/src/api/media_api.dart
@@ -15,6 +15,7 @@ import 'package:wenyou_api/src/model/create_upload_url_dto.dart';
 import 'package:wenyou_api/src/model/media_confirm_upload200_response.dart';
 import 'package:wenyou_api/src/model/media_get_media200_response.dart';
 import 'package:wenyou_api/src/model/media_get_upload_url201_response.dart';
+import 'package:wenyou_api/src/model/media_reissue_upload_url200_response.dart';
 
 class MediaApi {
 
@@ -296,6 +297,87 @@ class MediaApi {
     }
 
     return Response<MediaGetUploadUrl201Response>(
+      data: _responseData,
+      headers: _response.headers,
+      isRedirect: _response.isRedirect,
+      requestOptions: _response.requestOptions,
+      redirects: _response.redirects,
+      statusCode: _response.statusCode,
+      statusMessage: _response.statusMessage,
+      extra: _response.extra,
+    );
+  }
+
+  /// 为同一 UPLOADING 媒体重新签发临时上传地址
+  ///
+  ///
+  /// Parameters:
+  /// * [id]
+  /// * [cancelToken] - A [CancelToken] that can be used to cancel the operation
+  /// * [headers] - Can be used to add additional headers to the request
+  /// * [extras] - Can be used to add flags to the request
+  /// * [validateStatus] - A [ValidateStatus] callback that can be used to determine request success based on the HTTP status of the response
+  /// * [onSendProgress] - A [ProgressCallback] that can be used to get the send progress
+  /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
+  ///
+  /// Returns a [Future] containing a [Response] with a [MediaReissueUploadUrl200Response] as data
+  /// Throws [DioException] if API call or serialization fails
+  Future<Response<MediaReissueUploadUrl200Response>> mediaReissueUploadUrl({
+    required String id,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) async {
+    final _path = r'/api/v1/media/{id}/upload-url'.replaceAll('{' r'id' '}', encodeQueryParameter(_serializers, id, const FullType(String)).toString());
+    final _options = Options(
+      method: r'POST',
+      headers: <String, dynamic>{
+        ...?headers,
+      },
+      extra: <String, dynamic>{
+        'secure': <Map<String, String>>[
+          {
+            'type': 'http',
+            'scheme': 'bearer',
+            'name': 'bearer',
+          },
+        ],
+        ...?extra,
+      },
+      validateStatus: validateStatus,
+    );
+
+    final _response = await _dio.request<Object>(
+      _path,
+      options: _options,
+      cancelToken: cancelToken,
+      onSendProgress: onSendProgress,
+      onReceiveProgress: onReceiveProgress,
+    );
+
+    MediaReissueUploadUrl200Response? _responseData;
+
+    try {
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(MediaReissueUploadUrl200Response),
+      ) as MediaReissueUploadUrl200Response;
+
+    } catch (error, stackTrace) {
+      throw DioException(
+        requestOptions: _response.requestOptions,
+        response: _response,
+        type: DioExceptionType.unknown,
+        error: error,
+        stackTrace: stackTrace,
+      );
+    }
+
+    return Response<MediaReissueUploadUrl200Response>(
       data: _responseData,
       headers: _response.headers,
       isRedirect: _response.isRedirect,