+++ packages/wenyou_api/lib/wenyou_api.dart
@@ -240,6 +240,7 @@ export 'package:wenyou_api/src/model/mark_direct_conversation_read_dto.dart';
 export 'package:wenyou_api/src/model/media_confirm_upload200_response.dart';
 export 'package:wenyou_api/src/model/media_get_media200_response.dart';
 export 'package:wenyou_api/src/model/media_get_upload_url201_response.dart';
+export 'package:wenyou_api/src/model/media_reissue_upload_url200_response.dart';
 export 'package:wenyou_api/src/model/media_response_dto.dart';
 export 'package:wenyou_api/src/model/mention_candidate_dto.dart';
 export 'package:wenyou_api/src/model/mention_candidates_response_dto.dart';